@@ -1,11 +1,7 @@
-﻿
-namespace MultiFactor.Ldap.Adapter.Configuration.Core
+﻿namespace MultiFactor.Ldap.Adapter.Configuration.Core;
+
+public interface IConfigurationProvider
 {
-    public interface IConfigurationProvider
-    {
-        public System.Configuration.Configuration GetRootConfiguration();
-
-        public System.Configuration.Configuration[] GetClientConfiguration();
-
-    }
+    public Config GetRootConfiguration();
+    public Config[] GetClientConfiguration();
 }
