@@ -1,4 +1,4 @@
-﻿using System.Text.RegularExpressions;
+﻿using System;
 
 namespace MultiFactor.Ldap.Adapter.Core.NameResolving.NameTranslators
 {
@@ -6,6 +6,16 @@ namespace MultiFactor.Ldap.Adapter.Core.NameResolving.NameTranslators
     {
         public string Translate(NameResolverContext nameTranslatorContext, string from)
         {
+            if (nameTranslatorContext is null)
+            {
+                throw new ArgumentNullException(nameof(nameTranslatorContext));
+            }
+
+            if (from is null)
+            {
+                throw new ArgumentNullException(nameof(from));
+            }
+
             if (nameTranslatorContext.Profile != null)
             {
                 return nameTranslatorContext.Profile.Upn;
