@@ -1,4 +1,5 @@
-﻿using System.Text.RegularExpressions;
+﻿using System;
+using System.Text.RegularExpressions;
 
 namespace MultiFactor.Ldap.Adapter.Core.NameResolving.NameTranslators
 {
@@ -6,16 +7,26 @@ namespace MultiFactor.Ldap.Adapter.Core.NameResolving.NameTranslators
     {
         public string Translate(NameResolverContext nameTranslatorContext, string from)
         {
+            if (nameTranslatorContext is null)
+            {
+                throw new ArgumentNullException(nameof(nameTranslatorContext));
+            }
+
+            if (from is null)
+            {
+                throw new ArgumentNullException(nameof(from));
+            }
+
             if (nameTranslatorContext.Profile != null)
             {
                 return nameTranslatorContext.Profile.Upn;
             }
             foreach(var domain in nameTranslatorContext.Domains)
             {
-                var regex = new Regex("@" + domain.NetbiosName.ToLower() + "$", RegexOptions.IgnoreCase);
+                var regex = new Regex($"@{domain.NetbiosName.ToLower()}$", RegexOptions.IgnoreCase);
                 if(regex.IsMatch(from))
                 {
-                    var result = regex.Replace(from, "@" + domain.Domain.ToLower());
+                    var result = regex.Replace(from, $"@{domain.Domain.ToLower()}");
                     return result;
                 }
             }
